Multi Model and GPU Uptime: Implement per-MLNode PoC tracking with model-based organization
Section 5: Implement per-MLNode PoC tracking with model-based organization

- Add double repeated MLNode array structure (`repeated ModelMLNodes`)
- Perfect model correspondence: `p.Models[i]` ↔ `p.MlNodes[i]`
- Per-MLNode PoC tracking with NodeId in PoCBatch
- Model-based MLNode distribution in setModelsForParticipants
- Enhanced ValidationWeight for model-indexed arrays
- Updated API layer for NodeNum → NodeId mapping

Closes: Section 5 tasks 5.1-5.10

+++ decentralized-api/broker/node_worker_commands.go
@@ -94,7 +94,7 @@ func (c StartPoCNodeCommand) Execute(ctx context.Context, worker *NodeWorker) No
 	// Start PoC
 	dto := mlnodeclient.BuildInitDto(
 		c.BlockHeight, c.PubKey, int64(c.TotalNodes),
-		int64(worker.node.Node.NodeNum), c.BlockHash, c.CallbackUrl,
+		worker.node.Node.NodeNum, c.BlockHash, c.CallbackUrl,
 	)
 	if err := worker.mlClient.InitGenerate(ctx, dto); err != nil {
 		logging.Error("Failed to start PoC", types.PoC, "node_id", worker.nodeId, "error", err)
@@ -158,7 +158,7 @@ func (c InitValidateNodeCommand) Execute(ctx context.Context, worker *NodeWorker
 
 	dto := mlnodeclient.BuildInitDto(
 		c.BlockHeight, c.PubKey, int64(c.TotalNodes),
-		int64(worker.node.Node.NodeNum), c.BlockHash, c.CallbackUrl,
+		worker.node.Node.NodeNum, c.BlockHash, c.CallbackUrl,
 	)
 
 	if err := worker.mlClient.InitValidate(ctx, dto); err != nil {
